applet/vpn-request: fix external UI mode
The applet agent should return all secrets that were passed to it even
if they were not meant to be asked to users, otherwise they are
dropped from the connection when the connection is modified through an
editor (e.g. nm-connection-editor).

Now we build a list of known secrets and populate the dialog with the
ones that must be asked. After the dialog quits we update the secret
values in the list and return all the secrets, including ones not
asked.

See https://gitlab.freedesktop.org/NetworkManager/NetworkManager/issues/193

Fixes: bce8d0a75c1d ("applet/vpn-request: add external UI mode")