applet-agent: Make sure VPN auth dialogs get any agent-owned secrets
For VPN profiles that explicitly set the agent-owned flag on any
secrets, the auth dialog needs to know about them so a) it can actually
use them and b) it can print them back out if they don't need to be
updated.  Failing to do this causes agent-owned secrets to be cleared
from the keyring during the secret update process.