COPYING: update GPL license text
This is the COPYING file from NetworkManager ([1]). It's basically
the same. The difference are some only punctuation, white space and
it refers to "Lesser" GPL (LGPL-2.1) instead of "Library" GPL (LGPL-2.0).

I think it's preferable to have the exact same license text in
network-manager-applet as in NetworkManager.

Also, this version is exactly the same we get from the GNU site ([2])
as of today.

The same was done for libnma ([3]).

[1] https://gitlab.freedesktop.org/NetworkManager/NetworkManager/blob/3ceb7bf099b61224290c036f19187c9c327ece84/COPYING
[2] https://www.gnu.org/licenses/old-licenses/gpl-2.0.txt
[3] https://gitlab.gnome.org/GNOME/libnma/-/merge_requests/10