version: bump version to 1.15.0 (in anticipation of 1.16.0 release)
For applet, we (usually) don't main stable branches upstream. Nor
did the versioning scheme so far (with 3 numbers) give space for
that.

Up to now, the first two numbers (1.8) indicate the minimum required
libnm version API. There is no particular good reason to do (or not
to do) that. It worked reasonably well.

However, let's change our versioning scheme. For one, we only seldom
bump the required libnm version, so expressing this in the version
number of applet is not spectacular. Also, we have other important
libraries too (gtk3/gtk4, libnma-1.8.27). It's not clear that libnm's
version number is more important than the one of other libraries
and why applet's version number should follow libnm's version number.

Start a new versioning scheme with the next release. The next release
will be 1.16.0 (still to reflect that we now require libnm 1.16.0).
Also, future (major) releases will bump the second version number. Basically,
we will adopt the same versioning scheme as NetworkManager, and their
versions will move independently.

Note that we are still in a development, pre-release version for 1.16.0.
Hence, bump the version to 1.15.0.