version: bump version to 1.8.99 (in anticipation of 1.16.0 release)
For applet, we (usually) don't main stable branches upstream. Nor
does the versioning scheme (with 3 numbers) give currently space for
that.

Up to now, the first two numbers (1.8) indicate the minimum requires
libnm version API. There is no particular good reason to do (or not
to do) that. However,  worked reasonably well.

We bumped the libnm version requirement to 1.16.0. Reflect that in
our version number.

Note that we are still in a development, pre-release version for 1.16.0.
Hence, bump the version to 1.8.99.