build: temporarily allow building against libnma 1.22
With the split of libnma to a new repository, we also moved the
"org.gnome.nm-applet.gschema.xml" file. For that (and maybe other
reasons), we really need to build master of applet against latest
libnma (1.8.28).

However, libnma is not yet widely packaged, so for development and testing
this is rather inconvenient. Relax the requirement to allow building against
older libnma.

This patch should eventually be reverted.