gitlab-ci: move logic from gitlab-ci.yml to script and install external libnma-1.8.28
I think a plain shell script makes it easier to implement the build
steps. It's easier to read and to extend.

Also, the gitlab-ci script tried to install libnma from rawhide.
However, while the libnma-1.8.26 package is currently built in koji,
it is still (for some reasons) not available for installation. The
build step would still install libnma-1.8.24 (which was still bundled
with applet). Aside that, there are currently other issues in rawhide
that prevent this from working.

This only worked accidentally, because the network-manager-applet build
does not yet explicitly require a new libnma package. However, in practice
it does already, because when building against libnma-2.8.24, the
"org.gnome.nm-applet.gschema.xml" is missing. Next we are going to explicitly
require libnma-1.8.28, so this wouldn't work anymore.

Instead I did a scratch build of libnma-1.8.28 in koji. Adjust the
build steps and install the package. This is a temporary solution,
that helps to boot strap the update to an unbundled libnma-1.8.28.