applet: fix crash in VPN secret dialog handling
If the VPN connection gets deactivated when the dialog is displayed,
the SecretsRequest is no longer valid. Referencing it in the dialog
callback causes a segmentation fault.

To fix this, disconnect the dialog handler when the secret request is
cancelled and install a new handler that just destroys the dialog.

https://bugzilla.redhat.com/show_bug.cgi?id=1775278