c-e: fix initializing drop-down list for connection.secondaries
It seems that the initialization of the drop-down list did not work
properly. I think that is due to the elements being sorted, and using
gtk_combo_box_set_active() to set the index. But then the insertion
index no longer corresponds to the sorted index that is shown in the list.

Avoid that by using gtk_combo_box_set_active_iter().

Also, the connection.secondaries is a list of UUIDs. It doesn't
guarantee that the UUID actually exists. If no suitable profile can be
found (to populate the drop-down list), add the UUID directly.

https://gitlab.gnome.org/GNOME/network-manager-applet/-/issues/103