8021x-security: Use NMAWs classes instead of WirelessSecurity
Convert the ethernet 8021x-security page to use the NMAWs8021x class
and remove the dependency on wireless-security to work around issue #92.
Take advantage of libnma having a pure 8021x ui (without the WPA-EAP
assumptions) meaning that the hack that adds the wireless setting is no
longer needed.  For some reason after fill_connection an empty
WIRELESS_SECURITY setting is still being added.

This dialog is leaking memory as much as it was before.

Also there's a seemingly unwanted "Private Key Passphrase" text field
added for methods like MD5, EAP-TTLS+MD5.. sometimes -- looks like it
doesn't get added when the MD5 method is selected for the first time
and the username/password are still empty.