applet: Work around appindicator "show"/"hide" signals quirk
The ubuntu version of libappindicator (also available on some other
distros like arch) doesn't emit the "show" and "hide" signals as the
code expects.  This results in issue #4.  Use one of the workarounds
documented in
https://bugs.launchpad.net/ubuntu/+source/libappindicator/+bug/522152 to
at least get a signal when the menu pops up and request a single Wi-Fi
scan to make up for the missing periodic scans.

This commit includes a fix by @thaller ([1]).

[1] https://gitlab.gnome.org/GNOME/network-manager-applet/-/merge_requests/89#note_964843

https://gitlab.gnome.org/GNOME/network-manager-applet/-/merge_requests/89