c-e: show nm-c-e in GNOME desktop
For a long time, nm-c-e was hidden in GNOME. That means, there
is not UI path to start the tool, which is undesirable.

On the other hand, gnome-control-center had a dependency on nm-c-e,
so usually you would end up having two network tools. That was deemed
confusing. As solution, nm-c-e was hidden.

We already wanted to do this ([1], [2], [3]), but then there was
pushback ([1], [4]).

The plan was to first make g-c-c independent of nm-c-e ([5], [6])
so that one could install GNOME/g-c-c without nm-c-e. A major part
of that work was splitting out libnma as a separate library ([7]).

We are now ready to do this and revert ([[4]) -- except for example
on Fedora g-c-c still "Recommends:" nm-c-c and nm-c-c gets installed
by default ([8]).

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1493508#c10
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1715143
[3] https://gitlab.gnome.org/GNOME/network-manager-applet/-/commit/24c5f40e729b0efc7c53f4b33ead006ab7e769a6
[4] https://gitlab.gnome.org/GNOME/network-manager-applet/-/commit/a0c31ed56caf894919f4f3381df59e9b63c05405
[5] https://gitlab.gnome.org/GNOME/gnome-control-center/-/issues/512
[6] https://gitlab.gnome.org/GNOME/gnome-control-center/-/issues/1157
[7] https://gitlab.gnome.org/GNOME/libnma
[8] https://bugzilla.redhat.com/show_bug.cgi?id=1887891#c6

This reverts commit a0c31ed56caf894919f4f3381df59e9b63c05405.