applet: use symbolic icons instead of our ones
This reworks the applet icon, so that doesn't look out of place in
modern desktop environments that utilize it, such as Xfce.

It does away with custom icons and the animation in favor of standard
symbolic icons provided by the desktop theme, making things more
consistent and simple along the way.