Wi-Fi/OWE: Fix Enhanced Open (OWE) not being saved properly from nm-connection-editor
When we create a Wi-Fi connection using Enhanced Open (OWE) Security
from nm-connection-editor and save it, it was not being saved and the
security was being set as "None", with Wireless Security Setting
being discarded. This was fixed.

This fix is temporary fix. It should (and will) be done from libnma,
OWE should be implemented in libnma. Meanwhile this gets done,
this solution handles this issue.

https://bugzilla.redhat.com/show_bug.cgi?id=2023156

Fixes: b85ed300781ce9f26df9cf8e659d1bc0bc173201

https://gitlab.gnome.org/GNOME/network-manager-applet/-/merge_requests/108