wifi: change description for wpa-psk to include also WPA3
Since NM 1.32, key-mgmt=wpa-psk also enables WPA3 transition
mode. Update the description in the applet and editor to reflect that
change.

For WPA Enterprise, WPA3 is basically the same as WPA2 but with PMF
required. Then we have a stronger WPA3 Enterprise mode enabled with
key-mgmt=wpa-eap-suite-b-192. Don't change "WPA/WPA2 Enterprise" to
mention "WPA3" as that could be confused with the stronger mode
(currently not supported by the editor and applet).

While at it, change descriptions to use "/" instead of "&" as I find
it more clear.

https://gitlab.gnome.org/GNOME/network-manager-applet/-/merge_requests/116