Revert "gitlab-ci: move logic from gitlab-ci.yml to script and install external libnma-1.8.28"
This reverts commit 240da3ea68caa9eac516d5c2bd993c0eb7cc58be.

It fails, because the temporary (famous last words) workaround that
pulls in new libnma rpms no longer works (tarball and repo gone).

Let's just undo all of this and just bump to Fedora 31 base instead.

There's little justification for use of a shell script, and I find it
difficult to follow.