Don't allow setting a gateway on a WireGuard connection
WireGuard needs a special routing table to ensure that only unmarked,
unencrypted packets are sent to WireGuard. Setting a gateway on the IPv4
or IPv6 settings tab creates a default route in the main routing table
which causes a routing loop. The simplest way to prevent that mistake is
to not display the Gateway column for WireGuard connections.

Fixes https://gitlab.gnome.org/GNOME/network-manager-applet/-/issues/213